+++ packages/logger/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rtcio/logger
 
+## 0.1.2
+
+### Patch Changes
+
+- dc60668: Remove colors from dependencies
+- 384bbd4: Fixed typescript build chain to output ESNext instead of commonjs
+
 ## 0.1.1
 
 ### Patch Changes