+++ packages/signaling/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @rtcio/signaling
 
+## 0.2.0
+
+### Minor Changes
+
+- dc60668: Add a new type of event to the SignalerEvents -- newSignalPeerConnected
+
+### Patch Changes
+
+- 384bbd4: Fixed typescript build chain to output ESNext instead of commonjs
+
 ## 0.1.1
 
 ### Patch Changes